Починил компиляцию клангом, форматирование, переключение режима по табу

+++ src/graphic_factory.h
@@ -1,17 +1,20 @@
 #pragma once
-#include <span>
 #include "graphic_obj.h"
+#include <span>
 
 class GraphicFactory {
 	std::span<GraphicBuilder *> ext_;
 	int i;
 
 public:
-	GraphicFactory(std::span<GraphicBuilder*> ext) : ext_{ext}, i{(int)ext.size() - 1} {}
+	GraphicFactory(std::span<GraphicBuilder *> ext)
+		: ext_{ext}, i{(int)ext.size() - 1}
+	{
+	}
 
 	~GraphicFactory() = default;
 
-	void make(int x, int y) const { ext_[i]->create(x, y); };
+	void make(int x, int y) const { ext_[(size_t)i]->create(x, y); };
 
 	void prev() noexcept
 	{
